Move timestream aggregator library code into socs
To facilitate testing it's helpful to be able to import code that supports an
Agent. Since Agents themselves can be pretty thin, we move the major supporting
class and associated functions into socs/agent/.
